Injector: Handle callback arguments by a ArgumentHandlingStrategy
ArgumentHandlingStrategy is STRICT when none of the arguments in the
handler method are annotated with @Arg.

This means that we should be like default @Inject, and throw an error
when all the methods from the target are not in the handler.

ArgumentHandlingStrategy is LIGHT when the handler method's arguments
are all annotated with @Arg.

This means that we don't care if they have all the arguments or not,
the point of @Arg is to allow them to only specify what they want.